feat: added functions and refactored neuron content
